created finetune to inference flow
