+++ src/inference/serve.sh
@@ -1,9 +1,18 @@
+#!/bin/bash
+
+if [ "$#" -ne 1 ]; then
+    echo "Usage: $0 <model_path>"
+    exit 1
+fi
+
+model_path="$1"
+
 echo 'Starting controller...'
 python -u -m fastchat.serve.controller --host 0.0.0.0 > ~/controller.log 2>&1 &
 sleep 10
 echo 'Starting model worker...'
 python -u -m fastchat.serve.model_worker \
-        --model-path "results/llama2/final_merged_model" --host 0.0.0.0 2>&1 \
+        --model-path "$model_path" --host 0.0.0.0 2>&1 \
         | tee model_worker.log &
 
 echo 'Waiting for model worker to start...'
